+++ android/src/main/kotlin/com/smileidentity/flutter/SmileIDPlugin.kt
@@ -73,7 +73,7 @@ class SmileIDPlugin :
         context = flutterPluginBinding.applicationContext
 
         // Set wrapper info for Flutter SDK
-        SmileID.setWrapperInfo(WrapperSdkName.Flutter, "11.2.10")
+        SmileID.setWrapperInfo(WrapperSdkName.Flutter, "11.2.11")
 
         flutterPluginBinding.platformViewRegistry.registerViewFactory(
             SmileIDBiometricKYC.VIEW_TYPE_ID,